Complete Stage 2 (Not just numbers)
Upgrade the program to work with words and lines in addition to numbers.
Program now accepts a command-line argument to control which type it reads and finds the largest/maximum.
Wrote several new functions keeping everything as generic as possible throughout. Only has two or three spots where we have to re-duplicate code.