Complete Stage 5 (Error_)
Change input flow so we now return a String[]. This was needed since the stage requires us to parse the input for valid long's if our -dataType argument is "long" and output a message if any are found. A side benefit is it made code simpler since we don't have to convert to an array for sorting, and Data.mapValues can take an array without changing the body of the method itself.

Main.java:
- update argument handling to now check for unrecognized options as well as no arguments for defined options

Data.java:
- Data.readFromScanner now returns a String[] instead of an ArrayList.
- Data.mapValues now accepts a generic array instead of ArrayList

Lines.java, Words.java:
- remove code to convert ArrayList<String> to String[]

Longs.java:
- add code to parse the returned String[] to a Long[] and output a message if any non-valid inputs are found